[partial desig] removed python3.10 feature
