+++ src/pycram/datastructures/partial_designator.py
@@ -1,11 +1,10 @@
 # used for delayed evaluation of typing until python 3.11 becomes mainstream
 from __future__ import annotations
 
-from collections.abc import Iterable
-
 from typing_extensions import Type, List, Tuple, Any, Dict, TYPE_CHECKING
 from itertools import product
 from inspect import signature
+from ..utils import _is_iterable
 
 
 if TYPE_CHECKING:
@@ -63,7 +62,7 @@ class PartialDesignator:
         return self
         #return PartialDesignator(self.performable, *fargs, **newkeywors)
 
-    def __iter__(self) -> ActionDesignatorDescription.Action:
+    def __iter__(self) -> ActionDescription:
         """
         Iterates over all possible permutations of the arguments and keyword arguments and creates a new performable
         object for each permutation. In case there are conflicting parameters the args will be used over the keyword
@@ -81,7 +80,7 @@ class PartialDesignator:
         :yields: A list with a possible permutation of the given arguments
         """
         self._init_kwargs()
-        iter_list = [x if PartialDesignator._is_iterable(x) else [x] for x in self.kwargs.values()]
+        iter_list = [x if _is_iterable(x) else [x] for x in self.kwargs.values()]
         for combination in product(*iter_list):
             yield combination
 
@@ -91,22 +90,8 @@ class PartialDesignator:
         automatically be generated by dataclasses. Therefore, this method is only useable on dataclasses.
         """
         if not self.kwargs:
-            self.kwargs = {arg: self.__dict__[arg] for arg in self.__match_args__}
+            self.kwargs = {arg: self.__dict__[arg] for arg in list(signature(self.__class__).parameters.keys())}
             self.performable = self.__class__
-
-    @staticmethod
-    def _is_iterable(obj: Any) -> bool:
-        """
-        Checks if the given object is iterable.
-
-        :param obj: The object that should be checked
-        :return: True if the object is iterable, False otherwise
-        """
-        try:
-            iter(obj)
-        except TypeError:
-            return False
-        return True
 
     def missing_parameter(self) -> List[str]:
         """
